refactor: fetch_blended to fetch_relation_rows
